pyproject.toml
[project]
name = "vibebolt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "docker>=7.1.0",
    "mcp[cli]>=1.6.0",
    "openai-agents>=0.0.13",
    "ruff>=0.11.6",
    "tenacity>=9.1.2",
]
